Update cumulus cavi ui to include products list
Java swing UI was modified to include a list of products
instead of a fixed number of checkboxes representing the
available products.  NetBeans was used to modify the UI
and then that code was ported over to the Jython script.
The Jython script has code allowing to be used in the
CAVI, calls itself, and runs outside the CWMS/RTS env.